Added rspec tests for openlldp facts.
-Added fixtures to stub lldptool example output.
-Fixed lib/facter/openlldp.rb to use Facter::Util::Resolution.which
instead of File.exists.  This allows us to not have to hardcode the path
to lldptool. (yay)
-Writting rspec tests for facts is a horrendous endeavor.  Little
documentation exists and not being a Rubyist exacerbates things.  I
still don't know how to test for code that is not executed (see contexts
'if lldptool is not installed' and 'with no ethernet interfaces').
There is only so much one can grasp by looking at examples.