Five more attempts at fixing Travis-CI failures.
I have made a discovery.  Never run unit tests that check for the
existence of a file where that file (/usr/sbin/lldptool) exists on your
workstation but not on Travis-CI.  Once you remove the file you may
discover that your understanding of the rspec tests was incorrect.

I am stuck using five "before :each do" that mostly repeat because I
cannot seem to get things to work in the nested contexts.